Abstract
At the end of 2019, an unprecedented pandemic in human history originated in Wuhan, China, caused by a coronavirus known as SARS-CoV-2. This pandemic has claimed the lives of millions of people worldwide and has significantly impacted the economies of nations and the daily lives of their citizens. Cuba was not exempt from this situation, experiencing major waves of infections and deaths, primarily attributed to the emergence of new virus variants.

To mitigate the impact and progression of the epidemic in Cuba, the country launched mass vaccination campaigns using domestically developed vaccines. This study employs a hierarchical Bayesian model, simulated using Stan through the R software, to estimate the key parameters of the SARS-CoV-2 epidemic in Cuba. These parameters are explained based on the emergence of new virus variants and the number of vaccinated individuals, allowing for the assessment of the impact of the vaccination campaign and the introduction of new variants on the evolution of the epidemic in Cuba.

The main impacts of the vaccination campaign in Cuba included a reduction in the proportion of susceptible individuals from 99% to 49%, revealing an estimated vaccine effectiveness of 50%. Similarly, the infection fatality rate (IFR) decreased from 1.4% \[1.29%–1.49%] to 0.62% \[0.499%–0.751%], reaching a minimum of 0.24% \[0.190%–0.308%]. Compared to the original Wuhan strain, the Delta, Beta, and Omicron variants proved to be more virulent, with Omicron being the most transmissible. In contrast, the Epsilon and Alpha variants were less transmissible than the Wuhan strain. Population-wide vaccination significantly contributed to reducing the IFR, while the Omicron variant increased the IFR compared to other variants that affected the Cuban population.